# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/voting/current": {
      "get": {
        "operationId": "AdminVotingController_getCurrentCycle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get current active voting cycle with contestant tallies",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/voting/cycles": {
      "get": {
        "operationId": "AdminVotingController_listCycles",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all voting cycles (paginated)",
        "tags": [
          "Admin — Voting Management"
        ]
      },
      "post": {
        "operationId": "AdminVotingController_startNewCycle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCycleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Start a new voting cycle (auto-closes any active cycle)",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/voting/cycles/{id}": {
      "get": {
        "operationId": "AdminVotingController_getCycleDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get specific voting cycle details with vote stats",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/voting/cycles/{id}/close": {
      "patch": {
        "operationId": "AdminVotingController_closeCycle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Manually close a voting cycle",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/voting/eliminate": {
      "post": {
        "operationId": "AdminVotingController_eliminateContestant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EliminateContestantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Eliminate a contestant from the current voting cycle",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/voting/contestants/{id}/history": {
      "get": {
        "operationId": "AdminVotingController_getContestantVoteHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Vote history for a specific contestant across all cycles",
        "tags": [
          "Admin — Voting Management"
        ]
      }
    },
    "/api/admin/contestants": {
      "get": {
        "operationId": "AdminContestantsController_listContestants",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Stage 1",
                "Stage 2",
                "Stage 3",
                "Stage 4",
                "Final"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "active",
                "eliminated",
                "winner"
              ],
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all contestants with filters, text search, pagination, and stage summary stats",
        "tags": [
          "Admin — Contestants Management"
        ]
      },
      "post": {
        "operationId": "AdminContestantsController_createContestant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContestantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new contestant directly from admin dashboard",
        "tags": [
          "Admin — Contestants Management"
        ]
      }
    },
    "/api/admin/contestants/{id}": {
      "get": {
        "operationId": "AdminContestantsController_getContestantDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get contestant full details by ID",
        "tags": [
          "Admin — Contestants Management"
        ]
      },
      "patch": {
        "operationId": "AdminContestantsController_updateContestant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContestantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update basic contestant details (name, country, bio, image)",
        "tags": [
          "Admin — Contestants Management"
        ]
      }
    },
    "/api/admin/contestants/{id}/stage": {
      "patch": {
        "operationId": "AdminContestantsController_moveContestantStage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveContestantStageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Move a contestant to a new stage (Stage 1 to Final) with stage history logging",
        "tags": [
          "Admin — Contestants Management"
        ]
      }
    },
    "/api/admin/contestants/{id}/status": {
      "patch": {
        "operationId": "AdminContestantsController_updateContestantStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContestantStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update contestant status (pending, active, eliminated, winner)",
        "tags": [
          "Admin — Contestants Management"
        ]
      }
    },
    "/api/admin/contestants/{id}/history": {
      "get": {
        "operationId": "AdminContestantsController_getContestantStageHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get stage movement history log for a contestant",
        "tags": [
          "Admin — Contestants Management"
        ]
      }
    }
  },
  "info": {
    "title": "Africa By Road API",
    "description": "Backend API for Africa By Road - Road trip community platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "StartCycleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name for the voting cycle (e.g. \"Week 3 Voting\", \"Semi-Finals\")",
            "example": "Week 3 Voting"
          }
        },
        "required": [
          "name"
        ]
      },
      "EliminateContestantDto": {
        "type": "object",
        "properties": {
          "contestantId": {
            "type": "string",
            "description": "The ID of the contestant to eliminate",
            "example": "60d21b4667d0d8992e610c85"
          },
          "reason": {
            "type": "string",
            "description": "Optional reason for the elimination",
            "example": "Lowest votes in Week 3 cycle"
          }
        },
        "required": [
          "contestantId"
        ]
      },
      "CreateContestantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the contestant",
            "example": "Amara Diallo"
          },
          "country": {
            "type": "string",
            "description": "Country represented by contestant",
            "example": "Senegal"
          },
          "bio": {
            "type": "string",
            "description": "Contestant biography / pitch",
            "example": "Passionate road trip adventurer from Dakar."
          },
          "imageUrl": {
            "type": "string",
            "description": "Profile image URL",
            "example": "https://cdn.africabyroad.com/contestants/amara.jpg"
          },
          "currentStage": {
            "type": "string",
            "description": "Initial competition stage",
            "enum": [
              "Stage 1",
              "Stage 2",
              "Stage 3",
              "Stage 4",
              "Final"
            ],
            "default": "Stage 1"
          },
          "status": {
            "type": "string",
            "description": "Initial contestant status",
            "enum": [
              "pending",
              "active",
              "eliminated",
              "winner"
            ],
            "default": "active"
          }
        },
        "required": [
          "name",
          "country",
          "bio",
          "imageUrl"
        ]
      },
      "UpdateContestantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated contestant name",
            "example": "Amara Diallo"
          },
          "country": {
            "type": "string",
            "description": "Updated country",
            "example": "Senegal"
          },
          "bio": {
            "type": "string",
            "description": "Updated biography",
            "example": "Updated bio details."
          },
          "imageUrl": {
            "type": "string",
            "description": "Updated profile image URL",
            "example": "https://cdn.africabyroad.com/contestants/amara-v2.jpg"
          }
        }
      },
      "MoveContestantStageDto": {
        "type": "object",
        "properties": {
          "stage": {
            "type": "string",
            "description": "Target competition stage",
            "enum": [
              "Stage 1",
              "Stage 2",
              "Stage 3",
              "Stage 4",
              "Final"
            ],
            "example": "Stage 2"
          },
          "reason": {
            "type": "string",
            "description": "Reason for moving the contestant to this stage",
            "example": "Passed Stage 1 audition and voting criteria."
          }
        },
        "required": [
          "stage"
        ]
      },
      "UpdateContestantStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Updated contestant status",
            "enum": [
              "pending",
              "active",
              "eliminated",
              "winner"
            ],
            "example": "active"
          },
          "reason": {
            "type": "string",
            "description": "Reason for status update",
            "example": "Approved after identity verification."
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}